# src/public/openapi.yaml
openapi: 3.0.0
info:
  title: Code Parser API
  version: 1.0.0
servers:
  - url: http://localhost:3000

paths:

  /files:
    get:
      operationId: getFiles
      summary: Get the list of files in this project
      responses:
        '200':
          description: Successful
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '500':
          description: Internal server error

  /functions:
    get:
      operationId: getFunctions
      summary: Get the list of all functions in all files in this project
      responses:
        '200':
          description: Successful
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    fileName:
                      type: string
                    functions:
                      type: array
                      items:
                        type: object
                        properties:
                          functionName:
                            type: string
                          startByte:
                            type: number
                          endByte:
                            type: number
        '500':
          description: Internal server error

  /files/{fileName}:
    get:
      operationId: getFileContent
      summary: Get content or range of bytes from a specific file in this project
      parameters:
        - name: fileName
          in: path
          required: true
          schema:
            type: string
        - name: startByte
          in: query
          required: false
          schema:
            type: number
        - name: endByte
          in: query
          required: false
          schema:
            type: number
      responses:
        '200':
          description: Successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  fileName:
                    type: string
                  content:
                    type: string
        '400':
          description: Bad request (missing or incorrect parameters)
        '404':
          description: Not found (file not found)
        '500':
          description: Internal server error

    post:
      operationId: postNewFile
      summary: Create a new file in the project with specified content
      parameters:
        - in: path
          name: fileName
          required: true
          schema:
            type: string
          description: The name of the file to create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
              required:
                - content
            example:
              content: "Hello, world!"
      responses:
        '201':
          description: File created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Bad request
        '500':
          description: Internal server error

  /files/{fileName}/functions:
    get:
      operationId: getFunctionsInFile
      parameters:
      - name: fileName
        in: path
        required: true
        schema:
          type: string
      summary: Get the list of functions in a specified file in this project
      responses:
        '200':
          description: 'Successful'
          content:
            application/json:
              schema:
                type: object
                properties:
                  fileName:
                    type: string
                  functions:
                    type: array
                    items:
                      type: object
                      properties:
                        functionName:
                          type: string
                        startByte:
                          type: number
                        endByte:
                          type: number
        '400':
          description: Bad request (missing or incorrect parameters)
        '404':
          description: Not found (file not found)
        '500':
          description: Internal server error

  /files/{fileName}/functions/{functionName}:
    get:
      operationId: getFunctionContent
      summary: Get the content of a specific function in this project
      parameters:
        - name: fileName
          in: path
          required: true
          schema:
            type: string
        - name: functionName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Successful'
          content:
            application/json:
              schema:
                type: object
                properties:
                  fileName:
                    type: string
                  functionName:
                    type: string
                  content:
                    type: object
                    properties:
                      minimal:
                        description: minimal content for this function
                        type: string
                      full:
                        description: full content of function
                        type: string
                  startByte:
                    description: first byte of function location in the file
                    type: number
                  endByte:
                    description: last byte of function location in the file
                    type: number

  /run-command:
    post:
      operationId: runCommand
      summary: Run a command and stream the output
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                command:
                  type: string
              required:
                - command
            example:
              command: "npm test"
      responses:
        '200':
          description: Successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  exitCode:
                    type: number
                  stdout:
                    type: string
                  stderr:
                    type: string
        '400':
          description: Bad request (missing or incorrect parameters)
        '500':
          description: Internal server error

  /dependencies:
    get:
      operationId: getDependencies
      summary: List project dependencies
      responses:
        '200':
          description: Successful execution of yarn or npm command.
          content:
            application/json:
              schema:
                type: object
                properties:
                  exitCode:
                    type: number
                  stdout:
                    type: string
                  stderr:
                    type: string
        '400':
          description: Bad request (missing or incorrect parameters)
        '500':
          description: Internal server error
    post:
      operationId: manageDependencies
      summary: Add, remove or update project dependencies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                operation:
                  type: string
                packageName:
                  type: string
                version:
                  type: string
              required:
                - operation
                - packageName
            example:
              operation: "add"
              packageName: "@types/node"
              version: "^20.3"
      responses:
        '200':
          description: Successful execution of yarn or npm command.
          content:
            application/json:
              schema:
                type: object
                properties:
                  exitCode:
                    type: number
                  stdout:
                    type: string
                  stderr:
                    type: string
        '400':
          description: Bad request (missing or incorrect parameters)
        '500':
          description: Internal server error